requirements.txt
anndata==0.11.1
cellpose==3.0.11
imageio==2.36.0
matplotlib==3.10.0
numpy==2.2.1
openslide_python==1.4.1
ortools==9.3.10497
pandas==2.2.3
Pillow==11.0.0
POT==0.9.5
pytest==8.3.3
rpy2==3.5.11
scanpy==1.10.4
scikit_learn==1.6.0
scipy==1.14.1
seaborn==0.13.2
skimage==0.0
torch==2.5.1
torchvision==0.20.1
tqdm==4.66.5
